A lot of refactoring, and every tool now works correctly with multi selections
Everything now works as it should with multiselections, but there is a UI problem.
When attaching selected objects to the transform actor the transform actor becomes visible in the Outliner and the actors in the outliner are moved to be children of the Parent.  The Transform Actor also stays visible in the outliner after the operation completes. Its kinda ugly to look at it once you notice.